+++ src/test/java/com/github/alexisjehan/javanilla/lang/array/LongArraysTest.java
@@ -84,21 +84,38 @@ final class LongArraysTest {
 	}
 
 	@Test
-	void testAdd() {
-		assertThat(LongArrays.add(LongArrays.EMPTY, 0L)).containsExactly(0L);
+	@SuppressWarnings("deprecation")
+	void testAddLegacy() {
 		assertThat(LongArrays.add(LongArrays.of(1L, 2L, 3L), 0, 0L)).containsExactly(0L, 1L, 2L, 3L);
 		assertThat(LongArrays.add(LongArrays.of(1L, 2L, 3L), 1, 0L)).containsExactly(1L, 0L, 2L, 3L);
 		assertThat(LongArrays.add(LongArrays.of(1L, 2L, 3L), 2, 0L)).containsExactly(1L, 2L, 0L, 3L);
 		assertThat(LongArrays.add(LongArrays.of(1L, 2L, 3L), 3, 0L)).containsExactly(1L, 2L, 3L, 0L);
+	}
+
+	@Test
+	void testAdd() {
+		assertThat(LongArrays.add(LongArrays.EMPTY, 0L)).containsExactly(0L);
+		assertThat(LongArrays.add(LongArrays.of(1L, 2L, 3L), 0L, 0)).containsExactly(0L, 1L, 2L, 3L);
+		assertThat(LongArrays.add(LongArrays.of(1L, 2L, 3L), 0L, 1)).containsExactly(1L, 0L, 2L, 3L);
+		assertThat(LongArrays.add(LongArrays.of(1L, 2L, 3L), 0L, 2)).containsExactly(1L, 2L, 0L, 3L);
+		assertThat(LongArrays.add(LongArrays.of(1L, 2L, 3L), 0L, 3)).containsExactly(1L, 2L, 3L, 0L);
 		assertThat(LongArrays.add(LongArrays.of(1L, 2L, 3L), 0L)).containsExactly(1L, 2L, 3L, 0L);
+	}
+
+	@Test
+	@SuppressWarnings("deprecation")
+	void testAddInvalidLegacy() {
+		assertThatNullPointerException().isThrownBy(() -> LongArrays.add(null, 0, 0L));
+		assertThatIllegalArgumentException().isThrownBy(() -> LongArrays.add(LongArrays.of(VALUES), -1, 0L));
+		assertThatIllegalArgumentException().isThrownBy(() -> LongArrays.add(LongArrays.of(VALUES), 3, 0L));
 	}
 
 	@Test
 	void testAddInvalid() {
 		assertThatNullPointerException().isThrownBy(() -> LongArrays.add(null, 0L));
-		assertThatNullPointerException().isThrownBy(() -> LongArrays.add(null, 0, 0L));
-		assertThatIllegalArgumentException().isThrownBy(() -> LongArrays.add(LongArrays.of(VALUES), -1, 0L));
-		assertThatIllegalArgumentException().isThrownBy(() -> LongArrays.add(LongArrays.of(VALUES), 3, 0L));
+		assertThatNullPointerException().isThrownBy(() -> LongArrays.add(null, 0L, 0));
+		assertThatIllegalArgumentException().isThrownBy(() -> LongArrays.add(LongArrays.of(VALUES), 0L, -1));
+		assertThatIllegalArgumentException().isThrownBy(() -> LongArrays.add(LongArrays.of(VALUES), 0L, 3));
 	}
 
 	@Test